requirements.txt
fastapi
uvicorn
python-multipart
pymongo
python-dotenv
bcrypt
pyjwt
websockets
passlib